Wnt-core initial update

+++ network.js
@@ -50,6 +50,7 @@ var peer_events_buffer = [];
 var assocKnownPeers = {};
 var assocBlockedPeers = {};
 var exchangeRates = {};
+var bWatchingForLight = false;
 
 if (process.browser){ // browser
 	console.log("defining .on() on ws");
@@ -153,6 +154,7 @@ function sendErrorResponse(ws, tag, error) {
 	sendResponse(ws, tag, {error: error});
 }
 
+
 // if a 2nd identical request is issued before we receive a response to the 1st request, then:
 // 1. its responseHandler will be called too but no second request will be sent to the wire
 // 2. bReroutable flag must be the same
@@ -213,7 +215,34 @@ function sendRequest(ws, command, params, bReroutable, responseHandler){
 		};
 		sendMessage(ws, 'request', content);
 	}
+	return tag;
 }
+
+
+function deletePendingRequest(ws, tag){
+	if (ws && ws.assocPendingRequests && ws.assocPendingRequests[tag]){
+		var pendingRequest = ws.assocPendingRequests[tag];
+		clearTimeout(pendingRequest.reroute_timer);
+		clearTimeout(pendingRequest.cancel_timer);
+		delete ws.assocPendingRequests[tag];
+
+		// if the request was rerouted, cancel all other pending requests
+		if (assocReroutedConnectionsByTag[tag]){
+			assocReroutedConnectionsByTag[tag].forEach(function(client){
+				if (client.assocPendingRequests[tag]){
+					clearTimeout(client.assocPendingRequests[tag].reroute_timer);
+					clearTimeout(client.assocPendingRequests[tag].cancel_timer);
+					delete client.assocPendingRequests[tag];
+				}
+			});
+			delete assocReroutedConnectionsByTag[tag];
+		}
+		return true;
+	}else{
+		return false;
+	}
+}
+
 
 function handleResponse(ws, tag, response){
 	var pendingRequest = ws.assocPendingRequests[tag];
@@ -225,22 +254,8 @@ function handleResponse(ws, tag, response){
 			responseHandler(ws, pendingRequest.request, response);
 		});
 	});
-	
-	clearTimeout(pendingRequest.reroute_timer);
-	clearTimeout(pendingRequest.cancel_timer);
-	delete ws.assocPendingRequests[tag];
-	
-	// if the request was rerouted, cancel all other pending requests
-	if (assocReroutedConnectionsByTag[tag]){
-		assocReroutedConnectionsByTag[tag].forEach(function(client){
-			if (client.assocPendingRequests[tag]){
-				clearTimeout(client.assocPendingRequests[tag].reroute_timer);
-				clearTimeout(client.assocPendingRequests[tag].cancel_timer);
-				delete client.assocPendingRequests[tag];
-			}
-		});
-		delete assocReroutedConnectionsByTag[tag];
-	}
+
+	deletePendingRequest(ws, tag);
 }
 
 function cancelRequestsOnClosedConnection(ws){
@@ -502,6 +517,16 @@ function getPeerWebSocket(peer){
 			return wss.clients[i];
 	return null;
 }
+
+function getInboundDeviceWebSocket(device_address){
+	for (var i=0; i<wss.clients.length; i++){
+		if (wss.clients[i].device_address === device_address)
+			return wss.clients[i];
+	}
+	return null;
+}
+
+
 
 function findOutboundPeerOrConnect(url, onOpen){
 	if (!url)
@@ -849,9 +874,7 @@ function havePendingJointRequest(unit){
 function purgeJunkUnhandledJoints(){
 	if (bCatchingUp || Date.now() - coming_online_time < 3600*1000 || wss.clients.length === 0 && arrOutboundPeers.length === 0)
 		return;
-	db.query("DELETE FROM unhandled_joints WHERE creation_date < "+db.addTime("-1 HOUR"), function(){
-		db.query("DELETE FROM dependencies WHERE NOT EXISTS (SELECT * FROM unhandled_joints WHERE unhandled_joints.unit=dependencies.unit)");
-	});
+	joint_storage.purgeOldUnhandledJoints();
 }
 
 function purgeJointAndDependenciesAndNotifyPeers(objJoint, error, onDone){
@@ -920,13 +943,9 @@ function handleJoint(ws, objJoint, bSaved, callbacks){
 					callbacks.ifJointError(error);
 				//	throw Error(error);
 					unlock();
-					db.query(
-						"INSERT INTO known_bad_joints (joint, json, error) VALUES (?,?,?)", 
-						[objectHash.getJointHash(objJoint), JSON.stringify(objJoint), error],
-						function(){
-							delete assocUnitsInWork[unit];
-						}
-					);
+					joint_storage.saveKnownBadJoint(objJoint, error, function(){
+						delete assocUnitsInWork[unit];
+					});
 					if (ws)
 						writeEvent('invalid', ws.host);
 					if (objJoint.unsigned)
@@ -1142,7 +1161,11 @@ function handleSavedJoint(objJoint, creation_ts, peer){
 				sendErrorResult(ws, unit, error);
 		},
 		ifNeedHashTree: function(){
-			throw Error("handleSavedJoint: need hash tree");
+			console.log("handleSavedJoint "+objJoint.unit.unit+": need hash tree, will retry later");
+			setTimeout(function(){
+				handleSavedJoint(objJoint, creation_ts, peer);
+			}, 1000);
+		//	throw Error("handleSavedJoint "+objJoint.unit.unit+": need hash tree");
 		},
 		ifNeedParentUnits: function(arrMissingUnits){
 			db.query("SELECT 1 FROM archived_joints WHERE unit IN(?) LIMIT 1", [arrMissingUnits], function(rows){
@@ -1245,6 +1268,8 @@ function notifyWatchers(objJoint, source_ws){
 		return;
 	if (objJoint.ball) // already stable, light clients will require a proof
 		return;
+	if (!bWatchingForLight)
+		return;
 	// this is a new unstable joint, light clients will accept it without proof
 	db.query("SELECT peer FROM watched_light_addresses WHERE address IN(?)", [arrAddresses], function(rows){
 		if (rows.length === 0)
@@ -1268,7 +1293,7 @@ function notifyWatchersAboutStableJoints(mci){
 		unlock(); // we don't need to block writes, we requested the lock just to wait that the current write completes
 		notifyLocalWatchedAddressesAboutStableJoints(mci);
 		console.log("notifyWatchersAboutStableJoints "+mci);
-		if (mci <= 1)
+		if (mci <= 1 || !bWatchingForLight)
 			return;
 		storage.findLastBallMciOfMci(db, mci, function(last_ball_mci){
 			storage.findLastBallMciOfMci(db, mci-1, function(prev_last_ball_mci){
@@ -1317,10 +1342,12 @@ function notifyLocalWatchedAddressesAboutStableJoints(mci){
 	}
 	if (arrWatchedAddresses.length > 0)
 		db.query(
-			"SELECT unit FROM units CROSS JOIN unit_authors USING(unit) WHERE main_chain_index=? AND address IN(?) AND sequence='good' \n\
+			"SELECT unit FROM units CROSS JOIN unit_authors USING(unit) \n\
+			WHERE main_chain_index=? AND address IN("+arrWatchedAddresses.map(db.escape).join(', ')+") AND sequence='good' \n\
 			UNION \n\
-			SELECT unit FROM units CROSS JOIN outputs USING(unit) WHERE main_chain_index=? AND address IN(?) AND sequence='good'",
-			[mci, arrWatchedAddresses, mci, arrWatchedAddresses],
+			SELECT unit FROM units CROSS JOIN outputs USING(unit) \n\
+			WHERE main_chain_index=? AND address IN("+arrWatchedAddresses.map(db.escape).join(', ')+") AND sequence='good'",
+			[mci, mci],
 			handleRows
 		);
 	db.query(
@@ -1544,7 +1571,7 @@ function handleCatchupChain(ws, request, response){
 	}
 	var catchupChain = response;
 	console.log('received catchup chain from '+ws.peer);
-	catchup.processCatchupChain(catchupChain, ws.peer, {
+	catchup.processCatchupChain(catchupChain, ws.peer, request.params.witnesses, {
 		ifError: function(error){
 			bWaitingForCatchupChain = false;
 			sendError(ws, error);
@@ -1610,16 +1637,17 @@ function handleHashTree(ws, request, response){
 
 function waitTillHashTreeFullyProcessedAndRequestNext(ws){
 	setTimeout(function(){
-		db.query("SELECT COUNT(*) AS count FROM hash_tree_balls LEFT JOIN units USING(unit) WHERE units.unit IS NULL", function(rows){
-			if (rows[0].count <= 30){
+	//	db.query("SELECT COUNT(*) AS count FROM hash_tree_balls LEFT JOIN units USING(unit) WHERE units.unit IS NULL", function(rows){
+			var count = Object.keys(storage.assocHashTreeUnitsByBall).length;
+			if (count <= 30){
 				findNextPeer(ws, function(next_ws){
 					requestNextHashTree(next_ws);
 				});
 			}
 			else
 				waitTillHashTreeFullyProcessedAndRequestNext(ws);
-		});
-	}, 1000);
+	//	});
+	}, 100);
 }
 
 
@@ -1851,7 +1879,7 @@ function requestHistoryFor(arrUnits, arrAddresses, onDone){
 				console.log(response.error);
 				return onDone(response.error);
 			}
-			light.processHistory(response, {
+			light.processHistory(response, arrWitnesses, {
 				ifError: function(err){
 					sendError(ws, err);
 					onDone(err);
@@ -2097,7 +2125,7 @@ function handleJustsaying(ws, subject, body){
 			break;
 			
 		case 'my_url':
-			if (!body)
+			if (!ValidationUtils.isNonemptyString(body))
 				return;
 			var url = body;
 			if (ws.bOutbound) // ignore: if you are outbound, I already know your url
@@ -2258,6 +2286,7 @@ function handleJustsaying(ws, subject, body){
 			var address = body;
 			if (!ValidationUtils.isValidAddress(address))
 				return sendError(ws, "address not valid");
+			bWatchingForLight = true;
 			db.query("INSERT "+db.getIgnore()+" INTO watched_light_addresses (peer, address) VALUES (?,?)", [ws.peer, address], function(){
 				sendInfo(ws, "now watching "+address);
 				// check if we already have something on this address
@@ -2300,7 +2329,11 @@ function handleJustsaying(ws, subject, body){
 			if (!ws.bLoggingIn && !ws.bLoggedIn) // accept from hub only
 				return;
 			ws.close(1000, "my core is old");
-			throw Error("Mandatory upgrade required, please check the release notes at https://github.com/byteball/byteball/releases and upgrade.");
+			throw Error("Mandatory upgrade required, please check the release notes at https://github.com/worldnuqumoritytransporters/wnt/releases and upgrade.");
+			break;
+			
+		case 'custom':
+			eventBus.emit('custom_justsaying', ws, body);
 			break;
 	}
 }
@@ -2473,7 +2506,7 @@ function handleRequest(ws, tag, command, params){
 					[message_hash, message_string, objDeviceMessage.to],
 					function(){
 						// if the addressee is connected, deliver immediately
-						wss.clients.forEach(function(client){
+						wss.clients.concat(arrOutboundPeers).forEach(function(client){
 							if (client.device_address === objDeviceMessage.to && (!client.max_message_length || message_string.length <= client.max_message_length)) {
 								sendJustsaying(client, 'hub/message', {
 									message_hash: message_hash,
@@ -2551,6 +2584,7 @@ function handleRequest(ws, tag, command, params){
 					},
 					ifOk: function(objResponse){
 						sendResponse(ws, tag, objResponse);
+						bWatchingForLight = true;
 						if (params.addresses)
 							db.query(
 								"INSERT "+db.getIgnore()+" INTO watched_light_addresses (peer, address) VALUES "+
@@ -2808,6 +2842,10 @@ function handleRequest(ws, tag, command, params){
 				sendResponse(ws, tag, rows[0]);
 			});
 			break;
+			
+		case 'custom':
+			eventBus.emit('custom_request', ws, params,tag);
+		break;
 	}
 }
 
@@ -2816,7 +2854,7 @@ function onWebsocketMessage(message) {
 	var ws = this;
 	
 	if (ws.readyState !== ws.OPEN)
-		return;
+		return console.log("received a message on socket with ready state "+ws.readyState);
 	
 	console.log('RECEIVED '+(message.length > 1000 ? message.substr(0,1000)+'... ('+message.length+' chars)' : message)+' from '+ws.peer);
 	ws.last_ts = Date.now();
@@ -2910,9 +2948,11 @@ function startAcceptingConnections(){
 			tryHandleMessage();
 		});
 		ws.on('close', function(){
-			db.query("DELETE FROM watched_light_addresses WHERE peer=?", [ws.peer]);
-			db.query("DELETE FROM watched_light_units WHERE peer=?", [ws.peer]);
-			//db.query("DELETE FROM light_peer_witnesses WHERE peer=?", [ws.peer]);
+			if (bWatchingForLight){
+				db.query("DELETE FROM watched_light_addresses WHERE peer=?", [ws.peer]);
+				db.query("DELETE FROM watched_light_units WHERE peer=?", [ws.peer]);
+				//db.query("DELETE FROM light_peer_witnesses WHERE peer=?", [ws.peer]);
+			}
 			console.log("client "+ws.peer+" disconnected");
 			cancelRequestsOnClosedConnection(ws);
 		});
@@ -2932,6 +2972,7 @@ function startRelay(){
 		startAcceptingConnections();
 	
 	storage.initCaches();
+	joint_storage.initUnhandledAndKnownBad();
 	checkCatchupLeftovers();
 
 	if (conf.bWantNewPeers){
@@ -2996,6 +3037,7 @@ exports.sendJustsaying = sendJustsaying;
 exports.sendAllInboundJustsaying = sendAllInboundJustsaying;
 exports.sendError = sendError;
 exports.sendRequest = sendRequest;
+exports.sendResponse = sendResponse;
 exports.findOutboundPeerOrConnect = findOutboundPeerOrConnect;
 exports.handleOnlineJoint = handleOnlineJoint;
 
@@ -3020,3 +3062,5 @@ exports.isConnected = isConnected;
 exports.isCatchingUp = isCatchingUp;
 exports.requestHistoryFor = requestHistoryFor;
 exports.exchangeRates = exchangeRates;
+exports.getInboundDeviceWebSocket = getInboundDeviceWebSocket;
+exports.deletePendingRequest = deletePendingRequest;
